% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolong.R
\name{autolong}
\alias{autolong}
\title{Automatically aggregate, combine and harmonize multiple survey data sources.}
\usage{
autolong(info.csv)
}
\arguments{
\item{info.csv}{a csv file that contains the required codinf information}
}
\value{
A matrix of the infile
}
\description{
This function combines, aggregates, and harmonizes multiple datasets into a
single long format dataset. The user must provide a standardized csv table
containing all necessary coding information, which can be found in the
github repository.
}
